after corrupting db

+++ src/main/java/com/oss/ossv1/data/repository/ProductRepository.java
@@ -1,8 +1,8 @@
 package com.oss.ossv1.data.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.oss.ossv1.data.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 /**
  * This interface gives you methods like:
@@ -13,6 +13,6 @@ import com.oss.ossv1.data.entity.Product;
  *
  * You can use them directly in your ProductService or ProductController.
  */
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
     // You don't need to add anything here
 }
